refactor(components): relocate and restructure OnboardingTour to a dedicated directory
